fix: add get_token.py to the utilities.api_functions folder, remove get token from api_projects.py, edit api test
