+++ logic/api_client/api_projects.py
@@ -3,35 +3,20 @@ import json
 import requests
 
 from logic.api_client.data.body_data import ApiData as AD
-from utilities.regular_functions import reg_token
+from utilities.api_functions.get_token import Token
 
 
 class Projects:
     """ API library for website https://web.teambooktest.com """
 
     def __init__(self):
-        self.token = None
+        self.token = Token().get_token()
         self.base_url = AD.API_BASE_URL
-
-    def get_token(self) -> json:
-        """ Request to site swagger to get a users token using the specified email and password """
-        if self.token:
-            return self.token
-
-        try:
-            data = AD.login_body
-            res = requests.post(self.base_url + 'auth/login', data)
-            self.token = reg_token(res.text)
-            status = res.status_code
-            return status, self.token
-        except requests.exceptions.RequestException as e:
-            raise Exception(f"An error occurred while processing this request: {e}")
 
     def create_project(self) -> json:
         """Positive: Create a new project with required data"""
         try:
-            token = self.get_token()
-            params = {'token': token}
+            params = {'token': self.token}
             data = AD.create_project_required_data
             res = requests.post(self.base_url + 'projects', data=data, params=params)
             status = res.status_code
@@ -43,8 +28,7 @@ class Projects:
     def deactivate_project(self, project_id) -> json:
         """Positive: Deactivate a project"""
         try:
-            token = self.get_token()
-            params = {'token': token,
+            params = {'token': self.token,
                       'project_ids[]': project_id
                       }
             res = requests.patch(self.base_url + 'projects/deactivate', params=params)
@@ -56,9 +40,8 @@ class Projects:
     def activate_project(self, project_id) -> json:
         """Positive: Activate a project """
         try:
-            token = self.get_token()
             params = {
-                'token': token,
+                'token': self.token,
                 'project_ids[]': project_id
             }
             res = requests.patch(self.base_url + 'projects/activate', params=params)
@@ -70,9 +53,8 @@ class Projects:
     def delete_project(self, project_id) -> json:
         """Positive: Delete a project """
         try:
-            token = self.get_token()
             params = {
-                'token': token,
+                'token': self.token,
                 'project_ids[]': project_id
             }
             res = requests.patch(self.base_url + 'projects/delete', params=params)
@@ -82,7 +64,6 @@ class Projects:
             raise Exception(f"An error occurred while processing this request: {e}")
 
 
-# Projects().get_token()
 # Projects().create_project()
 # Projects().get_managers()
 # Projects().deactivate_project()
